+++ loggers/winston/index.js
@@ -45,13 +45,27 @@ const reservedFields = {
   res: true
 }
 
-// https://github.com/winstonjs/winston#creating-custom-formats
+// Create a Winston format for ecs-logging output.
+//
+// @param {Object} opts - Optional.
+//    - {Boolean} opts.convertErr - Whether to convert a logged `err` field
+//      to ECS error fields. Default true.
+//    - {Boolean} opts.convertReqRes - Whether to convert logged `req` and `res`
+//      HTTP request and response fields to ECS HTTP, User agent, and URL
+//      fields. Default false.
+//    - {Boolean} opts.apmIntegration - Whether to automatically integrate with
+//      Elastic APM (https://github.com/elastic/apm-agent-nodejs). If a started
+//      APM agent is detected, then log records will include the following
+//      fields:
+//        - "service.name" - the configured serviceName in the agent
+//        - "event.dataset" - set to "$serviceName.log" for correlation in Kibana
+//        - "trace.id", "transaction.id", and "span.id" - if there is a current
+//          active trace when the log call is made
+//      Default true.
 function ecsTransform (info, opts) {
-  // Boolean options for whether to specially handle some logged field names:
-  //  - `err` to ECS Error fields
-  //  - `req` and `res` to ECS HTTP, User agent, etc. fields
   let convertErr = true
   let convertReqRes = false
+  let apmIntegration = true
   // istanbul ignore else
   if (opts) {
     if (hasOwnProperty.call(opts, 'convertErr')) {
@@ -59,6 +73,9 @@ function ecsTransform (info, opts) {
     }
     if (hasOwnProperty.call(opts, 'convertReqRes')) {
       convertReqRes = opts.convertReqRes
+    }
+    if (hasOwnProperty.call(opts, 'apmIntegration')) {
+      apmIntegration = opts.apmIntegration
     }
   }
 
@@ -78,8 +95,10 @@ function ecsTransform (info, opts) {
     }
   }
 
-  // If there is a *started* APM agent, then use it.
-  const apm = elasticApm && elasticApm.isStarted && elasticApm.isStarted() ? elasticApm : null
+  let apm = null
+  if (apmIntegration && elasticApm && elasticApm.isStarted && elasticApm.isStarted()) {
+    apm = elasticApm
+  }
 
   // istanbul ignore else
   if (apm) {